+++ src/skipresult.rs
@@ -18,7 +18,7 @@ use crate::lines::ParsedLine;
 #[derive(Debug)]
 pub struct SkipResult<'a> {
   pub nextline: Option<ParsedLine<'a>>,
-  pub skip_end: usize,
+  pub skip_end: i64,
 }
 
 impl<'a> fmt::Display for SkipResult<'a> {
